% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{choose.cluster.BIC}
\alias{choose.cluster.BIC}
\title{Selects subspace closest to given variable (according to BIC)}
\usage{
choose.cluster.BIC(variable, pcas, numberClusters)
}
\arguments{
\item{variable}{variable to be assigned}

\item{pcas}{orthogonal basis for different subspaces}

\item{numberClusters}{number of subspaces (clusters)}
}
\value{
index number of subspace closest to variable
}
\description{
The most similar subspace is choosen based on BIC criterion
}
\keyword{internal}

