% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{pca.BIC}
\alias{pca.BIC}
\title{BIC for PCA, as given by Minka}
\usage{
pca.BIC(X, k)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{k}{number of principal components fitted}
}
\value{
BIC value of BIC criterion
}
\description{
Computes the value of BIC criterion for given data set and
number of factors.
}
\references{
Automatic choice of dimensionality for PCA, Thomas P. Minka
}
\keyword{internal}

