% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{pca.new.BIC.fast}
\alias{pca.new.BIC.fast}
\title{Penalized likelihood for PCA}
\usage{
pca.new.BIC.fast(X, k)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{k}{number of principal components fitted}
}
\value{
BIC value of BIC criterion
}
\description{
Computes the value of BIC-like criterion for given data set and
number of factors. Assumes that number of variables is large
compared to number of observations
}
\keyword{internal}

