% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pesel.R, R/pesel_package.R
\docType{package}
\name{pesel}
\alias{pesel}
\alias{pesel-package}
\title{Automatic estimation of number of principal components in PCA
with PEnalized SEmi-integrated Likelihood (PESEL)}
\usage{
pesel(X, npc.min = 1, npc.max = 10, scale = FALSE,
  method = c("heterogenous", "homogenous"), verbose = FALSE)
}
\arguments{
\item{X}{a data frame or a matrix contatining only continuous variables}

\item{npc.min}{minimal number of principal components, for all the possible
number of PCs between npc.min and npc.max criterion is computed}

\item{npc.max}{maximal number of principal components, if greater than
dimensions of X, min(ncol(X), nrow(X))-1 is used, for all the possible
number of PCs between npc.min and npc.max criterion is computed}

\item{scale}{a boolean, if TRUE (default value) then data is scaled before
applying criterion}

\item{method}{name of criterion to be used}

\item{verbose}{a boolean, if TRUE plot with BIC values for different
numbers of components is produced (default value is FALSE)}
}
\value{
number of components
}
\description{
Underlying assumption is that only small number of principal components,
associated with largest singular values, is relevent, while the rest of them
is noise. For a given numeric data set, function estimates the number of PCs
according to penalized likelihood criterion. Function adjusts the model used
to the case when number of variables is larger than the number of
observations.

pesel
}
\details{
Please note that no categorical variables and missing values are allowed.

Version: 0.3.0
}
\examples{
\dontrun{
library(MetabolAnalyze)
data(UrineSpectra)
pesel(UrineSpectra[[1]], method = "heterogenous", verbose=TRUE)
}
\dontrun{
library(varclust)
set.seeed(1)
sim.data <- data.simulation(n = 100, SNR = 1, K=1, numb.vars = 200, max.dim = 5)
pesel(sim.data$X, 1, 10)
}
}
\author{
Piotr Sobczyk,
        Julie Josse,
        Malgorzata Bogdan

Maintainer: Piotr Sobczyk \email{Piotr.Sobczyk@pwr.edu.pl}
}

