% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{pesel_heterogeneous_big_n}
\alias{pesel_heterogeneous_big_n}
\title{PEnalized SEmi-integrated Likelihood for heterogeneous singular values and
large number of observations}
\usage{
pesel_heterogeneous_big_n(X, minK, maxK)
}
\arguments{
\item{X}{a matrix containing only continuous variables}

\item{minK}{minimal number of principal components fitted}

\item{maxK}{maximal number of principal components fitted}
}
\value{
numeric vector, PESEL criterion for each k in range [minK, maxK]
}
\description{
Equivalent to BIC for PCA, as given by
Minka, \emph{Automatic choice of dimensionality for PCA}
}
\references{
Automatic choice of dimensionality for PCA, Thomas P. Minka
}
