% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{plot.mlcc.fit}
\alias{plot.mlcc.fit}
\title{Plot mlcc.fit class object}
\usage{
\method{plot}{mlcc.fit}(x, ...)
}
\arguments{
\item{x}{mlcc.fit class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Plot mlcc.fit class object
}
\keyword{internal}

