% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesel.R
\name{plot.pesel.result}
\alias{plot.pesel.result}
\title{Plot pesel.result class object}
\usage{
\method{plot}{pesel.result}(x, ...)
}
\arguments{
\item{x}{pesel.result class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Plot pesel.result class object
}
\keyword{internal}

