% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesel.R
\name{print.pesel.result}
\alias{print.pesel.result}
\title{Print pesel.result class object}
\usage{
\method{print}{pesel.result}(x, ...)
}
\arguments{
\item{x}{pesel.result class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Print pesel.result class object
}
\keyword{internal}

