% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxilliary.R
\name{rajan.noBIC}
\alias{rajan.noBIC}
\title{Non-penalized likelihood criterion for PCA based on paper by Rajan, Rayner}
\usage{
rajan.noBIC(X, k)
}
\arguments{
\item{X}{a matrix with only continuous variables}

\item{k}{number of principal components fitted}
}
\value{
BIC value of BIC criterion
}
\description{
Computes the value of BIC criterion for given data set and
number of factors. Assumes normal distribution of coefficients.
}
\references{
Automatic choice of dimensionality for PCA, Thomas P. Minka
}
\keyword{internal}

