% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pesel_package.R
\docType{package}
\name{varclust}
\alias{varclust}
\alias{varclust-package}
\title{PESEL}
\description{
PESEL
}
\details{
Version: 0.1.0
}
\author{
Piotr Sobczyk,
        Julie Josse,
        Malgorzata Bogdan

Maintainer: Piotr Sobczyk \email{Piotr.Sobczyk@pwr.edu.pl}
}

