% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pesel.R
\name{pesel}
\alias{pesel}
\title{Automatic estimation of number of principal components in PCA}
\usage{
pesel(X, npc.min = 1, npc.max = 10, scale = FALSE,
  method = c("heterogenous", "homogenous"), verbose = FALSE)
}
\arguments{
\item{X}{a data frame or a matrix contatining only continuous variables}

\item{npc.min}{minimal number of principal components, for all the possible number of
PCs between npc.min and npc.max criterion is computed}

\item{npc.max}{maximal number of principal components, if greater than dimensions of X,
min(ncol(X), nrow(X))-1 is used, for all the possible number of
PCs between npc.min and npc.max criterion is computed}

\item{scale}{a boolean, if TRUE (default value) then data is scaled before applying
criterion}

\item{verbose}{a boolean, if TRUE plot with BIC values for different
numbers of components is produced (default value is FALSE)}
}
\value{
number of components
}
\description{
Underlying assumption is that only small number of principal components, associated with largest singular values,
is relevent, while the rest of them is noise. For a given numeric data set, function estimates the number of PCs according to
penalized likelihood criterion. Function adjusts the model used to the case when number of variables is larger
than the number of observations.
}
\details{
Please note that no categorical variables and missing values are allowed.
}
\examples{
\dontrun{
library(MetabolAnalyze)
data(UrineSpectra)
pesel(UrineSpectra[[1]], method = "heterogenous", verbose=TRUE)
}
}

