% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesel.R
\name{plot.pesel.result}
\alias{plot.pesel.result}
\title{Plot pesel.result class object}
\usage{
\method{plot}{pesel.result}(x, posterior = TRUE, ...)
}
\arguments{
\item{x}{pesel.result class object}

\item{posterior}{a boolean, if TRUE (default value) then posterior probablities are plotted
otherwise values of PeSeL criterion are plotted}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
No return value, called for side effects
}
\description{
Plot pesel.result class object
}
\keyword{internal}
